Proposal of new Interface to handle complex task IComplexTask.
Proposal of Entity class to manage unit values such as Velocity, Attack Damage , Time to buld ... etc

The Drawing implementation does not contain hard-coding. Instead the bar is calulated by Current Health and MaxHealth.

CurrentHealth setter and MaxHelth setter is meant to call when a game is restored.However int he cose of the Resorces MaxHealth hasto be calc when loading the match.